requirements.txt
langchain
langchain-openai
langchain-pinecone
langchain-community
langchain-huggingface
pinecone
python-dotenv
sentence-transformers
pypdf